requirements_zor.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
mistralai>=1.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0
